chore: Improve release workflow
This commit improves the release workflow by using GitHub Actions' built-in actions for creating releases and uploading assets. This simplifies the
 workflow and makes it more reliable.

Specifically, the `actions/create-release` and `actions/upload-release-asset` actions are now used to create releases and upload APKs, respectively. This eliminates the need for manual curl commands and JSON parsing.

Signed-off-by: ViniciusDev <viniciusdev37@gmail.com>